A_BattMonitor: support resetting up to 32 batteries

+++ libraries/AP_BattMonitor/AP_BattMonitor.h
@@ -258,7 +258,7 @@ public:
     void checkPoweringOff(void);
 
     // reset battery remaining percentage
-    bool reset_remaining_mask(uint16_t battery_mask, float percentage);
+    bool reset_remaining_mask(uint32_t battery_mask, float percentage);
     bool reset_remaining(uint8_t instance, float percentage) { return reset_remaining_mask(1U<<instance, percentage);}
 
     // Returns mavlink charge state
